refactor pirinterface
